fix: rename plural word to singular and fix other mistakes
