+++ app/backend/src/application/workspace/get-workspace.ts
@@ -4,6 +4,6 @@ import WorkspaceSchema from 'database/models/workspace';
 /**
  * Returns stored workspace
  */
-export async function getWorkspace(): Promise<Workspace | any> {
+export async function getWorkspace(): Promise<Workspace | null> {
   return WorkspaceSchema.findOne();
 }
